+++ tests/UnitTests/TestTransfersConsumer.cpp
@@ -437,7 +437,7 @@ TEST_F(TransfersConsumerTest, onNewBlocks_getTransactionOutsGlobalIndicesError) 
   block.transactions.push_back(tx);
 
   consumer.addSubscription(subscription);
-  ASSERT_FALSE(consumer.onNewBlocks(&block, subscription.syncStart.height, 1));
+  ASSERT_FALSE(consumer.onNewBlocks(&block, static_cast<uint32_t>(subscription.syncStart.height), 1));
 }
 
 TEST_F(TransfersConsumerTest, onNewBlocks_updateHeight) {
@@ -457,7 +457,7 @@ TEST_F(TransfersConsumerTest, onNewBlocks_updateHeight) {
   block.block->timestamp = subscription.syncStart.timestamp;
   block.transactions.push_back(tx);
 
-  ASSERT_TRUE(m_consumer.onNewBlocks(&block, subscription.syncStart.height, 1));
+  ASSERT_TRUE(m_consumer.onNewBlocks(&block, static_cast<uint32_t>(subscription.syncStart.height), 1));
   ASSERT_EQ(900, container.balance(ITransfersContainer::IncludeAllLocked));
 
   std::unique_ptr<CompleteBlock[]> blocks(new CompleteBlock[subscription.transactionSpendableAge]);
@@ -468,7 +468,7 @@ TEST_F(TransfersConsumerTest, onNewBlocks_updateHeight) {
     addTestKeyOutput(*tr, 100, i + 1, generateAccountKeys());
   }
 
-  ASSERT_TRUE(m_consumer.onNewBlocks(blocks.get(), subscription.syncStart.height + 1, subscription.transactionSpendableAge));
+  ASSERT_TRUE(m_consumer.onNewBlocks(blocks.get(), static_cast<uint32_t>(subscription.syncStart.height + 1), static_cast<uint32_t>(subscription.transactionSpendableAge)));
   ASSERT_EQ(0, container.balance(ITransfersContainer::IncludeAllLocked));
   ASSERT_EQ(900, container.balance(ITransfersContainer::IncludeAllUnlocked));
 }
@@ -792,7 +792,7 @@ TEST_F(TransfersConsumerTest, onNewBlocks_manyBlocks) {
    }
  }
 
- ASSERT_TRUE(m_consumer.onNewBlocks(&blocks[0], 0, blocks.size()));
+ ASSERT_TRUE(m_consumer.onNewBlocks(&blocks[0], 0, static_cast<uint32_t>(blocks.size())));
 
  ASSERT_EQ(expectedTransactions, container.transactionsCount());
  ASSERT_EQ(expectedAmount, container.balance(ITransfersContainer::IncludeAll));
@@ -1031,7 +1031,7 @@ TEST_F(TransfersConsumerPerformanceTest, DISABLED_memory) {
 
   {
     AutoPrintTimer t;
-    ASSERT_TRUE(m_consumer.onNewBlocks(&blocks[0], 0, blocks.size()));
+    ASSERT_TRUE(m_consumer.onNewBlocks(&blocks[0], 0, static_cast<uint32_t>(blocks.size())));
   }
 
   blocks.clear();
@@ -1056,7 +1056,7 @@ TEST_F(TransfersConsumerPerformanceTest, DISABLED_performanceTest) {
   
   std::cout << "Calling onNewBlocks" << std::endl;
 
-  ASSERT_TRUE(m_consumer.onNewBlocks(&blocks[0], 0, blocks.size()));
+  ASSERT_TRUE(m_consumer.onNewBlocks(&blocks[0], 0, static_cast<uint32_t>(blocks.size())));
 
   auto end = std::chrono::steady_clock::now();
   std::chrono::duration<double> dur = end - start;
